Change size type to unsigned to cancel a warning

+++ bufferfd.h
@@ -5,7 +5,7 @@ class read_bufferfd : public autofd {
     static const size_t DEFAULT_BUF_SIZE;
     const size_t buf_size;
     auto_array<char> buffer;
-    mutable int startpos, endpos;
+    mutable size_t startpos, endpos;
 
     ssize_t buffer_copy( void *buf, size_t count ) const;
 public:
